test: config: allow for configurable image name
This change is to allow for specifying the name of the image to be used in tests via an env var - for use with boards where for testing we want to use a different image to the default "balena.img" export from the yocto build. E.g We use a flasher image for testing, but in production the user doesn't.

Change-type: patch
Signed-off-by: Ryan Cooke <ryan@balena.io>