initrdscripts: Allow passing extra kernel arguments to kexec
At this moment the kexec initrd script just takes the original kernel
command line, replaces root with UUID and removes bootloader args.

We have found at least one use-case (on the Pi4 and firmware GPIOs),
where a different initrd script needs to pass extra arguments
to the kexec'd kernel. With this patch it will append the contents
of the KEXEC_EXTRA_ARGS variable to the kernel command line.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>