initrdscripts: make the kexec script fail hard in unexpected states
At this moment the kexec initrd script is skipped when ROOTFS_DIR
is not defined or if the new rootfs is mounted, but does not contain
a kernel image in the expected place. This is undesirable as we
assume this is the last script executed by the balena bootloader.

This patch makes the kexec script always execute in the balena
bootloader and makes it fail hard in unexpected states, which means
the script is always an exit point for the balena bootloader,
whether the actual kexec call succeeds or not.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>