init: always run cryptsetup hook with secure boot
Init hooks specify an *_enabled function that determines whether or not
the hook is skipped during boot. Hooks are run by the "init" script,
shipped as part of the initramfs-framework poky recipe.

Several conditions were checked in the cryptsetup-efi-tpm enablement
function, some of which could cause the hook to be improperly skipped.

Force the hook to run when in secure boot user mode, as we want any
failure in the hook's run function to be detected and abort the boot.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>