tests: cloud: use local SSH auth to fetch logs during teardown
we are currently using SSH with balena cloud authentication to ssh into the DUT to fetch journal logs during the teardown. Its possible that the SSH keys registered to balena cloud are deleted by this point, causing auth failures and a very long set of retires at the end of the suite.

Change-type: patch
Signed-off-by: Ryan Cooke <ryan@balena.io>