update-balena-supervisor: add network-online dependency
There is a race condition present when purging a device and rebooting. If "update-balena-supervisor" comes up before the device can connect to the internet, it will fail to pull the supervisor image. As this service stops "balena-supervisor.service" , this isn't attempted again for at least 15 min = (up to 1 hour) when the "update-balena-supervisor.timer" triggers it.

Change-type: patch
Signed-off-by: Ryan Cooke <ryan@balena.io>