classes/image_types_balena: Add support for device specific boot filesystem options
The beagleplay is unable to load u-boot from the sd-card
boot partition if the standard geometry set by mkfs.vfat is used -
that is 16 heads / 63 sectors per track.

We noticed that using a geometry of 16 heads / 32 sectors per track
make the device to boot from the sd-card, so we make the
boot filesystem options configurable from the device repository
to apply this setting.

Change-type: patch
Signed-off-by: Alexandru Costache <alexandru@balena.io>