tests:os: Avoid setting apiEndpoint for unmanaged tests
When this config.json field has a value, the supervisor will treat the device as managed and will crash and start to boot loop.

On faster device-types we don't seem to notice and the tests pass by chance, but on Pi Zero and similar we consistently see ECONNRESET when hitting local SV endpoints.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>