os-helpers-fs: add support for specifying root partition by partuuid
This is currently unused as secure boot enabled devices are using the
partition labels given that the filesystem UUIDs are not readily
available for encrypted partitions.

However, a better solution that would avoid partition label clashes with
multiple partitions sharing labels is using partition UUIDs.

Other changes are needed to support this, but it makes sense to add the
general support for it to the helper functions.

Signed-off-by: Alex Gonzalez <alexg@balena.io>