initrdscripts: Wait for boot partition in the abroot script
At this moment the abroot script assumes that the boot partition
is already in place when it executes. This might not be true
if the rootfs sits on a device that takes a while to initialize,
such as a USB drive. The script fails hard if that is the case.

This patch replicates a waiting loop from the rootfs script,
which addresses the same issue for systems that do not use
the balena bootloader.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>