os-helpers-fs: introduce a script to split boot partitions
Secure boot enabled partitions need to split the boot partition into
an encrypted and a non-encrypted one.

Unless the device type natively separates the files that the bootROM
needs to boot, like EFI does into /mnt/boot/EFI, a list of files to
move to the non-encrypted partition needs to be provided on the
BALENA_NONENCRYPTED_BOOT_PARTITION_FILES variable, and a function
to implement the split needs to be provided.

This commit introduces a split_bootpartition() function that performs
this and can be used for those platforms that need to list the boot
essential files in a BALENA_NONENCRYPTED_BOOT_PARTITION_FILES variable.

This is typically called from the bootpart_split() function in the
balena-init-flasher customization file.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>