initrdscripts: migrate: panic on installation failure
Right now if the flashing script errors out the initramfs just keeps
running modules. This is a security risk specially for secure boot
systems as at that point we have an authorized trusted OS running in
an unvetted path.

This commit exits init if the flasher returns, and also attemps to crash
the kernel followed by an infinite sleep for paranoic reasons.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>