hostapp: Move supervisor update logic from balenahup
Right now the hostOS update is done through a call to the balenahup proxy action.
The code is https://github.com/balena-os/balenahup/blob/master/upgrade-2.x.sh

We are moving away from this OS update path over to having the supervisor handle OS updates.
As such, we need to move some of the steps from upgrade-2.x.sh over to the hostapp-update script.

For a time, until the supervisor managed OS updates are implemented, we will still be using OS updates
through the balenahup proxy action with the changes in this commit which should still work as before.

Signed-off-by: Florin Sarbu <florin@balena.io>
Change-type: patch