resin-init-flasher: with secure boot, authenticate the inner image
At this moment resin-init-flasher just takes whatever image lies in /opt
and dd's it to the target drive. This is fine for general use, but with
secure boot enabled, we want to perform at least basic authentication
of the image being written.

This patch gets the image signed at build time and makes flasher verify
the signature against a key built-in the kernel trust store. At this
very moment it fails hard if the signature does not match, but this may
change in the future. Technically we only want to know if we are about
to flash a balena-provided image or not, we might want to support both
but behave slightly differently in each scenario.

Change-type: minor
Signed-off-by: Michal Toman <michalt@balena.io>