usb-modeswitch: Update to version 2.6.1
Version 2.5.2 of usb-modeswitch contained a local 0.72 jimtcl copy which was used
when the static version of the dispatcher was being compiled. Version 2.6.1 removed
this copy of jimtcl from the source package so we add it back ourselves because it's
easier to have libjim.a compiled here rather than add a whole new Yocto recipe just for it.

Change-type: minor
Changelog-entry: Update usb-modeswitch to version 2.6.1
Signed-off-by: Florin Sarbu <florin@balena.io>