resin-mounts: only run non-encrypted mount if partition exists
When signed images are used without opting in secure boot, we don't want
the encrypted mount service to run and fail as there is no encrypted
boot partition.

This commit adds a condition to the encrypted boot mount service to only
run if the encrypted boot partition exists, similarly to what is
already done for EFI builds.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>