efitools: Allow enrolling raw hashes into UEFI trust store
At this moment hash-to-efi-sig-list only allows to enroll hashes
if the whole EFI application is presented, but this is technically
not necessary.

We want to be able to enroll raw hashes, because for PCIe devices
using 3rd party drivers, we only know the hash, the EFI application
or driver is not accessible to userspace.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>