resin-init-flasher: Support devices using 3rd party UEFI drivers
At this moment, if anything is measured into PCR2 (OpROMs, UEFI drivers)
the device will provision, but not be able to boot the first time.
This is because the EFI applications measured in PCR2 need to verify
against `db`, which means they will be loaded during provisioning
(while secure boot is still off), but won't be loaded once secure boot
is enabled. This changes the value of PCR2 between provisioning
and first boot, which means the encryption keys won't be released.

This patch adds two new behaviors to flasher:
* By default ignore the PCR2 hashes during secure boot provisioning,
  assuming none of the drivers will work with secure boot enabled.
* After an opt-in, read the hashes present in PCR2 and enroll
  them into `db` during provisioning.

The latter is an opt-in, as we have no control and no way to know
what a UEFI driver does. Since these run at firmware level, they have
enough privileges to bypass secure boot. All we can do is blindly
whitelist everything present during provisioning, which might be seen
as security risk, and the user should explicitly accept that.

Change-type: minor
Signed-off-by: Michal Toman <michalt@balena.io>