tests/device-tree: Rework test to not use the sysfs gpio interface
The /sys/class/gpio interface is bound to be removed anytime so let's
not use it anymore. Instead just make the test set the pin to high then
check its value after reboot then set the pin to low and check its value
again after the next reboot.

Also, when checking the pin value in /sys/kernel/debug/gpio lets search
for the gpio using the GPIO label instead of grepping for a gpio number
because various kernel versions may offset the gpios with a base.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>