Add DEVICE_TYPE
Allow for device type names and machine to differ. Until know, the
device type contract had to be named `$MACHINE.json`, and then the
slug in it could be something else.

The DEVICE_TYPE variable that defaults to MACHINE for backwards
compatibility allows for the same MACHINE to build several device types
by passing a different DEVICE_TYPE to the build.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>