meta-balena-YOCTO_RELEASE: Add compatibility with newer bmaptool
Yocto Scarthgap added bmaptool and also renamed it from bmap-tools. Now with the support
for Scarthgap that we added, because in meta-balena-common/classes/image_types_balena.bbclass
there is a dependency on this tool, we need to make sure we do not break the older Yocto
releases we still support and build.

Signed-off-by: Florin Sarbu <florin@balena.io>