balena-init-flasher-efi: look for kernel in both / and /tmp
One of the things that balena-init-flasher-efi does is that it copies
the currently running kernel to the EFI partition to serve as 2nd stage
bootloader when secure boot is enabled and the boot partition is split.
Right now it looks for the kernel in /, which works for provisioning
from an external drive, but not for the migrator. The migrator copies
the kernel into /tmp before running flasher, but that directory would
be skipped by flasher, as it is a different filesystem.

This patch makes flasher look for the kernel in both / and /tmp.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>