Replace the usage of /dev/fd/1 by /proc/self/fd/1
Currently we are using /dev/fd/1 as output file for tools that do not
normally allow using stdout. This works on a fully running system,
but the /dev/fd symlinks do not exist in the intramfs. This breaks
the migrator when provisioning secure boot systems.

This patch replaces /dev/fd/1 by /proc/self/fd/1 which is semantically
the same, but does exist even in the initramfs.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>