resin-init-flasher: Do not check exitcode of reboot, halt or poweroff
At this moment flasher tries multiple methods to shut down or reboot
the device after it has finished. The current approach first tries
a graceful way, e.g. `poweroff`, and checks the exitcode.
If the command exits with a non-zero status, it then adds a `-f` flag
to force the behavior. But this doesn't work inside the initramfs,
because the busybox variant of the commands exits with 0 and does
nothing (it will register the signal, but only shutdown / reboot
after the last initscript has been executed). Additionally,
the systemd variant of the tools, used on fully booted systems,
never returns to shell, so there is no point checking the exitcode.

This patch makes flasher use the forced variant if the graceful
attempt returned to shell, regardless of the exitcode.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>