resin-init-flasher: Sleep before SysRq to let shutdown properly execute if need be
There are cases when at the end of the flashing procedure the board will execute
"shutdown -h now" but sometimes the shutdown procedure is not immediate and the script will
continue with the next instruction which is to use SysRq and on some hardware this can happen
right when the previous shutdown command is still running in the background and the effect will
be that the board does not actually power off in the end (eth carrier is still on, the power
circuitry is still online and so on.)

To fix this we sleep for 5 seconds to give the system a chance to properly shutdown. If it has
not been shutdown in 5 seconds then most likely shutdown is broken so we force it via SysRq.

Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>