initrdscripts: mount securityfs in migrate script
At this moment securityfs is not mounted in the initramfs,
so when flasher is running from the migrate script, it won't find
the TPM event log. This is fine for most systems, but on systems
that have extra TPM events logged and measured, the expected PCR7
value will be computed incorrectly.

This patch mounts securityfs before running flasher from the migrate
script.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>