tests: whitelist extra PCR2 events by default
By introducing the option to whitelist or ignore extra events in PCR2,
we have changed the default behavior of the system. Instead of copying
the contents of PCR2 during provisioning, we now expect no drivers
to be loaded, unless there was an explicit opt-in through config.json.
This is the behavior that we've seen on physical hardware, however
qemu behaves differently - it loads the PXE driver for the network card
unconditionally, even if secure boot is enabled and the driver
has not been whitelisted.

This patch makes all the testsuites opt-in for the PCR2 whitelist,
which is just a workaround to restore the previous behavior and unblock
the automated testsuite.

A proper fix should either be in leviathan, or in QEMU, or implementing
a proper test for the PCR2 feature, that checks all the possible states.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>