os-helpers-tpm2: Always use password protection for TPM NVRAM writes
At this moment tpm_nvram_store_passphrase checks whether the expected
policy is currently met or not, and acts differently in each case.
If the policy doesn't match, the TPM write is protected by a one-time
password that is lost when the function exits.
If the policy matches, it will be used to protect the write directly.
The latter seems to be broken at least in some scenarios (retrying
HUP after a rollback).

This patch removes the special-casing and makes
tpm_nvram_store_passphrase always use the password protection. While
using the current PCR states directly is a nice optimization, always
using the password is more universal and makes the code simpler.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>