systemd: Add /sbin and /bin to $PATH for being passed to all executed processes
Enabling the usrmerge feature changed the PATH variable passed to the processes that systemd executes:
PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin

Notice it is missing /sbin and /bin. This is a problem in rollbacks for example which can execute in
older OSes that use split bin and usr instead of merged bin and usr and so will not be able to find
the binaries and scripts they need to call.

Changelog-entry: Make systemd pass /sbin and /bin in $PATH to all processes for compatibility with older OSes
Signed-off-by: Florin Sarbu <florin@balena.io>