linux-firmware: Move the iwlwifi firmware cleanup in meta-balena-common
The firmware we keep after this cleanup here should be supported from kernel 5.0 onwards
so it should be safe to move the cleanup code to meta-balena-common for Yocto versions
starting with Dunfell and up.
Also we increase the min firmware version to 36 for 8000C and 8265 and introduce the
min firmware version to 46 for 9260-th-b0-jf-b0.

Changelog-entry: Move iwlwifi linux-firmware cleanup to meta-balena-common and adjust the min fw version for several iwlwifi packages
Signed-off-by: Florin Sarbu <florin@balena.io>