initrdscripts: migrate: use an estimated compressed size for raw image
The memory check before a migration run was not taking into account
that the tmp directory is now compressed ram.

This change approximates the compressed size of the raw image which is
the biggest in the calculation. The rest of the file sizes are left as
is to provide some buffer for approximation failures and not to complicate
the logic excessively.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>