Temporarily remove BBAI 64 workflow
This device-type can not be built anymore because
the upstream beagleboard git repository has added
Anubis, which makes fetching of sources impossible
from Yocto builds.

We've asked for a github mirror from the upstream maintainer.
Until that becomes available, let's remove the workflow for
this device-type to unblock merging of PRs.

Change-type: patch
Signed-off-by: Alexandru Costache <alexandru@balena.io>