linux-firmware: add compatibility with earlier versions of compression
Earlier versions used "0" or "1" for FIRMWARE_COMPRESSION
to either enable or disable firmware compression. With the update
to 20250211, the semantics have changed and FIRMWARE_COMPRESSION
now expects either an empty string or the algorithm used to compress
the firmware (xz or zst).

This patch adds a small compatibility hack that translates
FIRMWARE_COMPRESSION = "1" to FIRMWARE_COMPRESSION = "xz",
and FIRMWARE_COMPRESSION = "0" to FIRMWARE_COMPRESSION = "", to add
the old behavior back and unwind the cyclic dependency when updating
meta-balena and device repos.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>