recipes-connectivity: skip the cloud ssh key checks earlier for "root" user
For scarthgap builds, there seems to be a problem when trying to ssh into a device locally using "root" , as we do during some tests, if the device is unmanaged. This is becuase the api endpoint and device api key aren't defined.

For some reason this wasn't a problem before

Change-type: patch
Signed-off-by: Ryan Cooke <ryan@balena.io>