Fail if IOMMU is disabled on secure boot systems
At this moment we provision secure boot systems without checking
whether the IOMMU is enabled. With the IOMMU disabled, all PCIe cards
have full DMA access and are able to retrieve the encryption keys
from memory. On some systems this logs an extra event into PCR7.

This patch makes the OS look for the event in the TPM event log
and refuse to provision or unlock the disks if the IOMMU is disabled.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>