update-balena-supervisor.service: Only start this service after we have internet connectivity
There is a board that is slower to get to internet connectivity over WiFi (slower interface
bring-up, slower initial DNS resolution). Because of this, in cases where we need to update
the supervisor just after boot (such as after clearing the data partiton by means of deleting
/mnt/data/remove_me_to_reset and after that rebooting the board) the update-balena-supervisor.service
will fail because internet connectivity is not yet established.

To overcome this we make update-balena-supervisor.service depend on a separate target we have defined solely
for the purpose of waiting for network connectivity. See
https://github.com/balena-os/meta-balena/commit/a6781d54c6edced457ad916cac3fec9c70d25a0a for details

Changelog-entry: Condition the start of update-balena-supervisor.service after internet connectivity
Change-type: patch
Signed-off-by: Florin Sarbu <florin@balena.io>