kernel-balena.bbclass: Fix zram failure when setting algorithm
The initramfs error addressed by this commit is:

    zramctl: /dev/zram1: failed to set algorithm: Invalid argument

zstd is the default algorithm currently used by the zram initramfs module,
which is installed for all device-types. Also, the crypto API provides
the compression algorithms used by the zram kernel module, so let's make
this available for all devices as built-in, so it is accessible in the
initramfs.

Change-type: patch
Signed-off-by: Alexandru Costache <alexandru@balena.io>