initrdscripts: separate mounting boot partition from abroot
Right now, abroot is the only script executed in the balena bootloader
so it does everything - find the boot partition, mount it, read boot
configuration, perform A/B selection, and unmount the boot partition.

Long-term, we want to introduce other bootloader scripts that read
the boot configuration and do other things, so we want to have
an "entry" script that finds and mounts the boot partition, so that
the subsequently executed scripts can just use it.

This patch splits the abroot script into two - the mountboot script
that finds and mounts the boot partition, and the abroot script,
that only performs tha A/B selection.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>